+++ src/components/ChannelPreviewCompact.js
@@ -24,7 +24,7 @@ export class ChannelPreviewCompact extends React.PureComponent {
   channelPreviewButton = React.createRef();
 
   onSelectChannel = () => {
-    this.props.setActiveChannel(this.props.channel);
+    this.props.setActiveChannel(this.props.channel, this.props.watchers);
     this.channelPreviewButton.current.blur();
   };
   render() {