+++ src/components/ChannelPreviewLastMessage.js
@@ -25,7 +25,7 @@ export class ChannelPreviewLastMessage extends PureComponent {
   channelPreviewButton = React.createRef();
 
   onSelectChannel = () => {
-    this.props.setActiveChannel(this.props.channel);
+    this.props.setActiveChannel(this.props.channel, this.props.watchers);
     this.channelPreviewButton.current.blur();
   };